refactor: group XML backend parsers in a subfolder
