+++ docling/backend/xml/uspto_backend.py
